Change sitebody cmp from constant to dynamic component.
Declare SiteBody as var instead of const.
Implement if conditionals to change return depending on view from
parent component, site.

Create MainView component as first view option for sitebody.
MainView (so far) consists of two sub-componenents: a body com-
ponent to display content and a menu component to change what
content will be displayed.
Style the main view so that the body and menu are arranged
horizontally, and take up as much space as is available.
Initialize view of the main view to be the blog.