Add click handler for buttons in main menu.
Following example set by
https://tylermcginnis.com/react-js-tutorial-pt-1-a-comprehensive-guide-to-building-apps-with-react-js-8ce321b125ba#.7qx3w1yrn

Create changeView function in MainView class - this will change
the main view's subview by directly altering the state.

Pass changeView to children components for menu and menu buttons.
Also pass the view state itself to the menu and menu buttons.
Create a handler for clicks on any of the menu options, which calls
changeView to update the subview of the main view to the selected
option.
Convert list items into button elements encased within list items.
This allows the handler to be successfully initiated upon click.
Create a variable appliedStyles which will compute styles based on
a match/non-match between the button and the current subview of the
main view: highlighted button if the button is for the current view,
non-highlighted if not.