Convert menu options into components and add click handler.
Following example set by
https://tylermcginnis.com/react-js-tutorial-pt-1-a-comprehensive-guide-to-building-apps-with-react-js-8ce321b125ba#.7qx3w1yrn

Create changeView function in Site class - this will change the
site's view by directly altering the state.

Convert the menu component into a dynamic componant (instead
of a constant), and create menu option component, such that
menu options are generated from the site's view options rather
than hardcoded elements.

Pass changeView to children components for site menu and menu
options. Also pass the view state itself to the site menu and
menu options.

Create click handler for clicks on any menu options, which calls
changeView to update the view of the site to the selected option.

Create a variable appliedStyles which will compute styles based
on a match/non-match between the option and the current site
view: highlighted option if the option is for the current view,
non-highlighted if not.